# TODO-PLAN/cloudbox-openapi.yaml
openapi: 3.0.3
info:
  title: CloudBox API
  version: 1.0.0
  description: 私有文件管理与分享系统 V1 REST API。
servers:
  - url: http://localhost:8090/api/v1
    description: Local development
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      operationId: register
      summary: 注册用户
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: 登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshToken
      summary: 刷新令牌
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: 刷新成功，同时轮换 Refresh Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: 退出登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          $ref: '#/components/responses/SuccessNoData'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /folders:
    post:
      tags: [Nodes]
      operationId: createFolder
      summary: 新建目录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFolderRequest'
      responses:
        '201':
          description: 新目录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeResponse'
        '409':
          $ref: '#/components/responses/Conflict'
  /files/upload:
    post:
      tags: [Files]
      operationId: uploadFile
      summary: 上传单个文件
      description: 文件大小上限由服务端配置，V1 建议 100 MiB。服务端必须流式写入。
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                parentId:
                  type: integer
                  format: int64
                  description: 省略表示根目录
                conflictPolicy:
                  $ref: '#/components/schemas/ConflictPolicy'
                file:
                  type: string
                  format: binary
            encoding:
              file:
                contentType: application/octet-stream
      responses:
        '201':
          description: 上传成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeResponse'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          description: 文件过大
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /files/{fileId}/content:
    get:
      tags: [Files]
      operationId: downloadFile
      summary: 下载文件
      description: 支持 Range 请求。完整响应为 200，范围响应为 206。
      parameters:
        - $ref: '#/components/parameters/FileId'
        - name: Range
          in: header
          schema:
            type: string
          example: bytes=0-1048575
      responses:
        '200':
          description: 完整文件
          headers:
            Content-Disposition:
              schema:
                type: string
            Accept-Ranges:
              schema:
                type: string
                example: bytes
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '206':
          description: 部分文件
          headers:
            Content-Range:
              required: true
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
  /files/{fileId}/preview:
    get:
      tags: [Files]
      operationId: previewFile
      summary: 内联预览文件
      description: V1 仅允许浏览器安全展示的图片、纯文本和 PDF；其他类型返回 415。
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '200':
          description: 内联文件内容，Content-Disposition 为 inline
          content:
            image/*:
              schema:
                type: string
                format: binary
            text/plain:
              schema:
                type: string
            application/pdf:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
        '415':
          description: 不支持预览该文件类型
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /public/shares/{shareCode}/files/{fileId}/content:
    get:
      tags: [Public Shares]
      operationId: downloadPublicSharedFile
      summary: 下载分享中的文件
      security:
        - ShareToken: []
        - {}
      parameters:
        - $ref: '#/components/parameters/ShareCode'
        - $ref: '#/components/parameters/FileId'
        - name: Range
          in: header
          schema:
            type: string
      responses:
        '200':
          description: 完整文件
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '206':
          description: 部分文件
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '403':
          $ref: '#/components/responses/Forbidden'
        '410':
          $ref: '#/components/responses/Gone'
  /admin/users:
    get:
      tags: [Admin]
      operationId: listUsers
      summary: 分页查询用户
      security:
        - BearerAuth: []
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
            maxLength: 100
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/UserStatus'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 用户分页列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPageResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/users/{userId}/status:
    patch:
      tags: [Admin]
      operationId: updateUserStatus
      summary: 启用或禁用用户
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserStatusRequest'
      responses:
        '200':
          description: 更新后的用户
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/users/{userId}/quota:
    patch:
      tags: [Admin]
      operationId: updateUserQuota
      summary: 调整用户存储配额
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateQuotaRequest'
      responses:
        '200':
          description: 更新后的存储用量
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorageUsageResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ShareToken:
      type: apiKey
      in: header
      name: X-Share-Token
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    FileId:
      name: fileId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
    ShareCode:
      name: shareCode
      in: path
      required: true
      schema:
        type: string
        minLength: 6
        maxLength: 32
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
  responses:
    SuccessNoData:
      description: 操作成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseVoid'
    BadRequest:
      description: 请求参数或操作不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 未认证或凭据无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: 无权限访问
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: 资源状态或名称冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Gone:
      description: 分享已过期、取消或达到下载上限
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: 存储配额不足
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: 请求过于频繁
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          pattern: '^[A-Za-z0-9_]{4,32}$'
          example: alice_01
        email:
          type: string
          format: email
          maxLength: 128
          example: alice@example.com
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 72
          example: StrongPass123!
    LoginRequest:
      type: object
      required: [account, password]
      properties:
        account:
          type: string
          description: 用户名或邮箱
          maxLength: 128
        password:
          type: string
          format: password
          maxLength: 72
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          minLength: 20
...
    CreateFolderRequest:
      type: object
      required: [name]
      properties:
        parentId:
          type: integer
          format: int64
          nullable: true
          description: NULL 或省略表示根目录
        name:
          $ref: '#/components/schemas/NodeName'
...
    UpdateUserStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/UserStatus'
    UpdateQuotaRequest:
      type: object
      required: [quotaBytes]
      properties:
        quotaBytes:
          type: integer
          format: int64
          minimum: 0
    NodeName:
      type: string
      minLength: 1
      maxLength: 255
      pattern: '^(?![.]{1,2}$)(?!.*[\\/:*?"<>|]).+$'
      description: 不能为 . 或 ..，不能包含 Windows/Linux 保留分隔字符
...
    ConflictPolicy:
      type: string
      enum: [REJECT, RENAME]
      default: REJECT
    NodeType:
      type: string
      enum: [FILE, FOLDER]
    NodeStatus:
      type: string
      enum: [ACTIVE, TRASHED]
    UserStatus:
      type: string
      enum: [ACTIVE, DISABLED]
    UserRole:
      type: string
      enum: [USER, ADMIN]
...
    User:
      type: object
      required: [id, username, email, role, status, quotaBytes, usedBytes, createdAt]
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/UserStatus'
        quotaBytes:
          type: integer
          format: int64
        usedBytes:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
    TokenData:
      type: object
      required: [accessToken, refreshToken, tokenType, expiresIn, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          description: Access Token 剩余秒数
          example: 1800
        user:
          $ref: '#/components/schemas/User'
    StorageUsage:
      type: object
      required: [quotaBytes, usedBytes, availableBytes, usageRatio]
      properties:
        quotaBytes:
          type: integer
          format: int64
        usedBytes:
          type: integer
          format: int64
        availableBytes:
          type: integer
          format: int64
          minimum: 0
        usageRatio:
          type: number
          format: double
          minimum: 0
          example: 0.34
...
    Node:
      type: object
      required: [id, type, name, sizeBytes, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int64
        parentId:
          type: integer
          format: int64
          nullable: true
        type:
          $ref: '#/components/schemas/NodeType'
        name:
          type: string
        sizeBytes:
          type: integer
          format: int64
        contentType:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/NodeStatus'
        deletedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PageUser:
      type: object
      required: [items, page, size, total, pages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/User'
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer
          format: int64
        pages:
          type: integer
    ApiResponseVoid:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: OK
        message:
          type: string
          example: success
        data:
          nullable: true
          example: null
    UserResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: string
          example: OK
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/User'
    TokenResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: string
          example: OK
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/TokenData'
    StorageUsageResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: string
          example: OK
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/StorageUsage'
    NodeResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: string
          example: OK
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/Node'
...
    UserPageResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: string
          example: OK
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/PageUser'
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    ErrorResponse:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          type: string
          example: FILE_NAME_CONFLICT
        message:
          type: string
          example: 目标目录中已存在同名文件
        requestId:
          type: string
        errors:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/ErrorDetail'